+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [1.1.1] - 2023-09-17
+
 ### Changed
 
 - Player inventories, ender chests, and XP are now cleared on reset
@@ -40,7 +42,8 @@
 - Support for Paper 1.19.3+
 - English translation
 
-[unreleased]: https://github.com/qixils/fahare/compare/v1.1.0...HEAD
+[unreleased]: https://github.com/qixils/fahare/compare/v1.1.1...HEAD
+[1.1.1]: https://github.com/qixils/fahare/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/qixils/fahare/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/qixils/fahare/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/qixils/fahare/releases/tag/v1.0.0